[LLDB] Initial implementation of Data Inspection Language.
*** THIS IS NOT READY FOR FULL OR COMPLETE REVIEW YET ***

This is the intial pass at implementing the Data Inspection Language,
as described in
https://discourse.llvm.org/t/rfc-data-inspection-language/69893/3

Eventually this will be a replacement for the implementation of 'frame
variable'.  This initial implementation does not yet fully implement all
the features in the RFC; it is meant to only be equivalent to the current
'frame variable' command (although the initial implementation does do a
bit more).

This adds a new option to the 'frame var' command ('--dil') that will invoke
this implementation instead of the standard frame variable implementation.

This is still a work in progress, as it does not quite pass all the tests that
the current frame variable command passes.